// src/types/youtube.ts
export type LangId = string;

export type AdaptiveFormatRange = {
  end: string;
  start: string;
};

export type ColorInfo = {
  matrixCoefficients: "COLOR_MATRIX_COEFFICIENTS_BT709";
  primaries: "COLOR_PRIMARIES_BT709";
  transferCharacteristics: "COLOR_TRANSFER_CHARACTERISTICS_BT709";
};

export type Quality =
  | "tiny"
  | "small"
  | "medium"
  | "large"
  | "hd720"
  | "hd1080"
  | "hd1440"
  | "hd2160";

export type QualityLabel =
  | "144p"
  | "240p"
  | "360p"
  | "480p"
  | "720p"
  | "720p60"
  | "1080p"
  | "1080p60"
  | "1440p"
  | "1440p60"
  | "2160p60";

export type BaseAdaptiveFormat = {
  approxDurationMs: string;
  contentLength: string;
  averageBitrate: number;
  bitrate: number;
  indexRange: AdaptiveFormatRange;
  initRange: AdaptiveFormatRange;
  itag: number;
  lastModified: string;
  mimeType: string;
  projectionType: "RECTANGULAR";
  quality: Quality;
};

export type BaseAdaptiveFormatWithUrl = BaseAdaptiveFormat & {
  url: string;
};

export type BaseAdaptiveFormatWithSignatureCipher = BaseAdaptiveFormat & {
  signatureCipher: string;
};

export type UniversalAdaptiveFormat =
  | BaseAdaptiveFormatWithUrl
  | BaseAdaptiveFormatWithSignatureCipher;

export type VideoAdaptiveFormat = {
  colorInfo?: ColorInfo;
  fps: number;
  height: number;
  qualityLabel: QualityLabel;
  width: number;
} & UniversalAdaptiveFormat;

export type AdaptiveFormatAudioTrackInfo = {
  /**
   * Default audio by interface locale
   */
  audioIsDefault: boolean;
  /**
   * localized name of audio track
   */
  displayName: string;
  id: LangId;
};

export type AudioAdaptiveFormat = {
  audioChannels: number;
  audioQuality: "AUDIO_QUALITY_MEDIUM" | "AUDIO_QUALITY_LOW";
  audioSampleRate: string;
  /**
   * It's rarely. If you need you can find this in every MrBeast video
   */
  audioTrack?: AdaptiveFormatAudioTrackInfo;
  highReplication: boolean;
  loudnessDb: number;
  isDrc?: boolean;
  xtags?: string;
} & UniversalAdaptiveFormat;

export type AdaptiveFormat = VideoAdaptiveFormat | AudioAdaptiveFormat;
// ...
export type AudioTrackPlaceholder = {
  captionTrackIndices: number[];
};

export type CaptionInitialState =
  | "CAPTIONS_INITIAL_STATE_ON_RECOMMENDED"
  | "CAPTIONS_INITIAL_STATE_OFF_RECOMMENDED";

export type AudioTrackVisibility = "ON" | "UNKNOWN";

export interface AudioTrack extends AudioTrackPlaceholder {
  audioTrackId: LangId;
  captionTrackIndices: number[];
  captionsInitialState: CaptionInitialState;
  defaultCaptionTrackIndex: number;
  hasDefaultTrack: boolean;
  visibility: AudioTrackVisibility;
}